RFCT(Extracted test helper methods into BaseAutomationTests)
